image_processing/requirements.txt
Flask==2.0.2
gunicorn==20.1.0
requests==2.26.0
Pillow==8.3.2